move run from main to __init__.py (msk) and other files needed - WORKING
