+++ error_log.txt
@@ -1,1 +1,4 @@
 2024-11-13 19:36:34: test error
+2024-11-15 12:44:55: 'analyses'
+2024-11-15 12:47:17: 'analyses'
+2024-11-15 12:49:11:
