+++ example_data/example_scripts/setup_project_bops.py
