--- src/bops/settings.xml
@@ -1,71 +1,0 @@
-<settings>
-    <project>
-        <current_project_folder>/path/to/project</current_project_folder>
-        <type>biomechanics</type>
-        <filters>
-            <filter1>value1</filter1>
-            <filter2>value2</filter2>
-        </filters>
-        <subjects>
-            <subject>
-                <name>subject1</name>
-                <age>25</age>
-                <gender>male</gender>
-            </subject>
-            <subject>
-                <name>subject2</name>
-                <age>30</age>
-                <gender>female</gender>
-            </subject>
-        </subjects>
-        <analyses>
-            <processed>
-                <scale>true</scale>
-                <IK>true</IK>
-                <ID>true</ID>
-                <SO>true</SO>
-            </processed>
-        </analyses>
-    </project>
-    <paths>
-        <main>C:\Git\research_data\Projects\test_project</main>
-        <setup_folder>C:\Git\research_data\Projects\test_project\Setups</setup_folder>
-        <setup_ceinms>C:\Git\research_data\Projects\test_project\Setups\ceinms</setup_ceinms>
-        <simulations>C:\Git\research_data\Projects\test_project\Simulations</simulations>
-        <current_analysis>C:\Git\python-envs\msk_modelling\Lib\site-packages\msk_modelling_python\src\current_analysis.json</current_analysis>
-        <subject>C:\Git\research_data\Projects\test_project\Simulations\P01</subject>
-        <trial>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l</trial>
-        <results>C:\Git\research_data\Projects\test_project\results</results>
-        <c3d>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\c3dfile.c3d</c3d>
-        <grf>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\grf.mot</grf>
-        <markers>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\marker_experimental.trc</markers>
-        <emg>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\EMG_filtered.sto</emg>
-        <models>C:\Git\research_data\Projects\test_project\Scaled_models</models>
-        <model_generic>C:\Git\research_data\Projects\test_project\Scaled_models\generic_model.osim</model_generic>
-        <model_scaled>C:\Git\research_data\Projects\test_project\Scaled_models\P01_scaled.osim</model_scaled>
-        <grf_xml>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\GRF.xml</grf_xml>
-        <ik_setup>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\setup_ik.xml</ik_setup>
-        <id_setup>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\setup_id.xml</id_setup>
-        <ma_setup>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\setup_ma.xml</ma_setup>
-        <ik_output>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\IK.mot</ik_output>
-        <id_output>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\inverse_dynamics.sto</id_output>
-        <ma_output_folder>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\muscle_analysis</ma_output_folder>
-        <so_output_forces>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\_StaticOptimization_force.sto</so_output_forces>
-        <so_output_activations>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\_StaticOptimization_activation.sto</so_output_activations>
-        <so_actuators>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\actuators_so.xml</so_actuators>
-        <jra_output>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\joint_reaction.sto</jra_output>
-        <jra_setup>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\setup_jra.xml</jra_setup>
-        <ceinms_src>C:\Git\msk_modelling_matlab\src\Ceinms\CEINMS_2</ceinms_src>
-        <uncalibrated_subject>C:\Git\research_data\Projects\test_project\Simulations\P01\ceinms_shared\ceinms_uncalibrated_subject.xml</uncalibrated_subject>
-        <calibrated_subject>C:\Git\research_data\Projects\test_project\Simulations\P01\ceinms_shared\ceinms_calibrated_subject.xml</calibrated_subject>
-        <ceinms_exc_generator>C:\Git\research_data\Projects\test_project\Simulations\P01\ceinms_shared\ceinms_excitation_generator.xml</ceinms_exc_generator>
-        <ceinms_calibration_setup>C:\Git\research_data\Projects\test_project\Simulations\P01\ceinms_shared\ceinms_calibration_setup.xml</ceinms_calibration_setup>
-        <ceinms_trial_exe>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\ceinms_trial.xml</ceinms_trial_exe>
-        <ceinms_trial_cal>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\ceinms_trial_cal.xml</ceinms_trial_cal>
-        <ceinms_exe_setup>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\ceinms_exe_setup.xml</ceinms_exe_setup>
-        <ceinms_exe_cfg>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\ceinms_exe_cfg.xml</ceinms_exe_cfg>
-        <ceinms_results>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\ceinms_results</ceinms_results>
-        <ceinms_results_forces>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\ceinms_results\MuscleForces.sto</ceinms_results_forces>
-        <ceinms_results_activations>C:\Git\research_data\Projects\test_project\Simulations\P01\pre\dynamic04_l\ceinms_results\Activations.sto</ceinms_results_activations>
-    </paths>
-</settings>
